requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-pinecone
langchain-huggingface
langgraph
pinecone
transformers
torch
sentence-transformers
pypdf
pytest
python-dotenv
